# static/swagger.yaml
openapi: 3.0.3
info:
  title: md-to-pdf Document Engine API
  description: |
    Moteur de documents complet : conversion Markdown/HTML/Template vers PDF,
    thèmes, graphiques et diagrammes, aperçu PNG, fusion, filigrane, protection
    par mot de passe, caviardage, audit de mise en page et comparaison visuelle.

    ## Comportement conditionnel des réponses

    La plupart des endpoints de génération PDF acceptent les champs optionnels
    `client_id` et `pdf_name`. Selon leur présence :

    | `client_id` + `pdf_name` | Réponse |
    |--------------------------|---------|
    | **fournis** | Le PDF est sauvegardé sur le serveur et un JSON `{ "download_url": "/download/…" }` est retourné |
    | **absents** | Le fichier PDF binaire est retourné directement dans le body |

    `client_id` et `pdf_name` servent à construire un chemin sur disque : ils doivent être
    des noms simples (`[A-Za-z0-9._-]`, 128 caractères max, ne commençant pas par un point).

    ## Deux façons de désigner un PDF

    Partout où un corps de requête attend un PDF — `pdf`, `before`, `after`, les entrées de
    `pdfs` — **deux formes sont acceptées** :

    | Référence | Origine |
    |-----------|---------|
    | `/download/{client_id}/{nom}.pdf` | un PDF que ce service a produit et sauvegardé |
    | `asset://as_…` | un fichier déposé par l'appelant via `POST /api/files` |

    La première n'a pas changé et ne changera pas. La seconde est ce qui rend les outils
...

    ## Sortie des outils

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - ApiKeyHeader: []
  - BearerToken: []
...
paths:
...
  /api/redact:
    post:
      tags: [Redact]
      summary: Caviarder un PDF sauvegardé
      description: |
        Localise le texte à masquer, pose des rectangles noirs, puis **reconstruit chaque
        page à partir de pixels**. Le texte caviardé disparaît du fichier, et avec lui la
        couche texte, les signets, les annotations, les pièces jointes et les métadonnées
        de la source. Le résultat n'est ni sélectionnable ni cherchable, et il est plus
        lourd que l'original.

        Deux vérifications sont faites sur le fichier avant de le servir : le nombre de
        pages doit être celui de la source, et l'extraction de texte doit ne rien rendre.

        Au moins un de `patterns` ou `entities` est obligatoire.

        > Un PDF **scanné** (sans couche texte) ressort aplati avec `redactions: []` :
        > rien n'a pu être localisé, il n'y a pas d'OCR.
      operationId: redact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedactRequest"
            examples:
              entities:
                summary: Caviarder les données personnelles
                value:
                  pdf: "/download/client-123/contrat.pdf"
                  entities: [email, iban, phone]
                  client_id: "client-123"
                  pdf_name: "contrat-caviarde"
      responses:
        "200":
          description: PDF binaire, ou JSON quand `client_id` et `pdf_name` sont fournis
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                $ref: "#/components/schemas/RedactResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "504":
          $ref: "#/components/responses/Timeout"

  # ───────────────────────────────────────────────
  # Layout
  # ───────────────────────────────────────────────
...
  /api/compress:
    post:
      tags: [Compress]
      summary: Compresser un PDF, et dire ce que ça a coûté
      description: |
        Ghostscript en `-dSAFER`, quatre niveaux, et la facture attachée : le `verdict`
        compare le nombre de pages et le nombre de caractères extractibles avant et après.
        Un scan poussé en `screen` revient quatre fois plus léger et n'est plus cherchable ;
        c'est exactement ce que le verdict nomme.

        Une compression qui produit un fichier **plus gros** est une compression ratée : la
        source est alors renvoyée telle quelle, et le verdict le dit.
      operationId: compress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompressRequest"
            example:
              pdf: "asset://as_9f2c4b1d8e6a03f75c21b40d9e8a1c37"
              level: ebook
              output: asset
      responses:
        "200":
          $ref: "#/components/responses/ToolResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "504":
          $ref: "#/components/responses/Timeout"

  # ───────────────────────────────────────────────
  # Repair
  # ───────────────────────────────────────────────
...
  /api/unlock:
    post:
      tags: [Unlock]
      summary: Retirer une protection dont vous avez le mot de passe
      description: |
        `qpdf --decrypt` avec le mot de passe **fourni**, et rien d'autre. Cet endpoint ne
        cherche jamais un mot de passe : un `password` vide est refusé en 400. C'est la
        ligne entre un outil légitime et un outil de contournement, et elle est tracée
        volontairement.

        Le mot de passe passe par un fichier d'arguments : il n'apparaît jamais dans la
        liste des processus de la machine.
      operationId: unlock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnlockRequest"
            example:
              pdf: "asset://as_9f2c4b1d8e6a03f75c21b40d9e8a1c37"
              password: "secret123"
      responses:
        "200":
          $ref: "#/components/responses/ToolResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "504":
          $ref: "#/components/responses/Timeout"

  # ───────────────────────────────────────────────
  # Rasterize  (PDF → PNG / JPEG)
  # ───────────────────────────────────────────────
  /api/rasterize:
    post:
      tags: [Rasterize]
      summary: PDF → images
      description: |
        `pdftoppm` sur un document que le service n'a pas rendu lui-même — l'équivalent de
        `/api/preview` pour un fichier déposé.

        Une seule page suit les trois destinations habituelles. **Plusieurs pages répondent
        toujours en JSON**, une entrée par page dans `assets` : plusieurs images ne peuvent
        pas partager une URL de téléchargement, et l'avertissement le dit quand `client_id`
        et `pdf_name` ont été ignorés.

        Le plafond de pages par appel est `PDF_PREVIEW_MAX_PAGES` (20 par défaut) ; au-delà
        la réponse est écourtée et `warnings` le signale. Un plafond de pixels borne aussi
        la combinaison d'un dpi élevé et de nombreuses pages.
      operationId: rasterize
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RasterizeRequest"
            example:
              pdf: "asset://as_9f2c4b1d8e6a03f75c21b40d9e8a1c37"
              pages: "1-4"
              dpi: 150
              format: png
      responses:
        "200":
          description: |
            L'image produite, ou le JSON `ToolResponse` quand une destination a été demandée
            ou que plusieurs pages ont été rendues.
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                $ref: "#/components/schemas/ToolResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "504":
          $ref: "#/components/responses/Timeout"

  # ───────────────────────────────────────────────
  # Images → PDF
  # ───────────────────────────────────────────────
...
  /api/ocr:
    post:
      tags: [OCR]
      summary: Donner une couche texte à un document numérisé
      description: |
        `ocrmypdf`, et le sidecar relu page par page : une page qui est ressortie vide est
        **nommée** dans le verdict au lieu de voyager silencieusement dans un fichier qui a
        l'air correct.

        Tout ce qui suit en dépend : `/api/redact` rend `redactions: []` et `/api/extract`
        un document vide quand les pages sont des images de texte.

        Plafond de 300 pages par requête (et jamais plus que `ASSET_MAX_PAGES`) : au-delà,
        découper avec `/api/pages` puis traiter chaque partie. Vu le coût, préférer
        `POST /api/jobs`.
      operationId: ocr
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OcrRequest"
            example:
              pdf: "asset://as_9f2c4b1d8e6a03f75c21b40d9e8a1c37"
              languages: [fra, eng]
              mode: auto
              pdfa: false
      responses:
        "200":
          $ref: "#/components/responses/ToolResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "504":
          $ref: "#/components/responses/Timeout"

  # ───────────────────────────────────────────────
  # Extract  (PDF → Markdown / texte / JSON)
  # ───────────────────────────────────────────────
...
components:
...
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Clé d'API, active uniquement si `API_KEY` est définie côté serveur.
    BearerToken:
      type: http
      scheme: bearer
      description: |
        Même clé que `X-API-Key`, transmise via `Authorization: Bearer <clé>`.
    SessionCookie:
      type: apiKey
      in: cookie
      name: mdpdf_session
      description: |
        Cookie de session posé par `POST /api/auth/signup` et `POST /api/auth/login`.
        `HttpOnly`, `SameSite=Lax`, et `Secure` dès que `X-Forwarded-Proto` vaut `https`.

        Il ne se mélange jamais avec `X-API-Key` : le cookie dit « un humain regarde une
        page », l'en-tête dit « un programme appelle ». C'est cette séparation qui permet au
        cookie de rester `Lax` — donc de refuser de voyager sur un POST cross-site — pendant
        que la clé reste utilisable depuis n'importe où.

  # ─── Paramètres réutilisables ─────────────────
...
  responses:
...
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Bad request"
            details: "\"data\" must be a JSON object"
...